+++ .eslintrc.js
@@ -1,0 +1,8 @@
+module.exports = {
+    "extends": "google",
+    "rules": {
+        "max-len": 0,
+        "max-nested-callbacks": ["warn", {"max": 5}],
+        "valid-jsdoc": 0
+    }
+};