+++ bootstrap.js
@@ -5,8 +5,7 @@ module.exports = function(grunt) {
     grunt.initConfig({
       config: config
     });
-  }
-  else {
+  } else {
     grunt.config.set('config', config);
   }
 
@@ -16,10 +15,11 @@ module.exports = function(grunt) {
   // Wrap Grunt's loadNpmTasks() function to allow loading Grunt task modules
   // that are dependencies of Grunt Drupal Tasks.
   grunt._loadNpmTasks = grunt.loadNpmTasks;
-  grunt.loadNpmTasks = function (mod) {
+  grunt.loadNpmTasks = function(mod) {
     var internalMod = grunt.file.exists(__dirname, 'node_modules', mod);
+    var pathOrig;
     if (internalMod) {
-      var pathOrig = process.cwd();
+      pathOrig = process.cwd();
       process.chdir(__dirname);
     }
     grunt._loadNpmTasks(mod);
@@ -29,38 +29,46 @@ module.exports = function(grunt) {
   };
 
   // Load all tasks from grunt-drupal-tasks.
-  grunt.loadTasks(__dirname + '/tasks');
+  var path = require('path');
+  grunt.loadTasks(path.join(__dirname, '/tasks'));
 
   // Define the default task to fully build and configure the project.
   var tasksDefault = [];
 
-  // If the "--no-validate" option is given, skip adding "validate" to task array.
+  // If the "--no-validate" option is given, skip adding "validate" to default
+  // tasks array.
   if (!grunt.option('no-validate')) {
     tasksDefault.push('validate');
   }
 
-  // If build/html exists, but is empty, skip the newer check.
-  // This facilitates situations where the build/html is generated as a mounted
-  // directory point with a newer timestamp than the Drush Makefiles.
-  //
-  // We do not use the grunt-newer .cache with drushmake so skipping newer for
-  // any one run does not impact later behavior.
-  if (grunt.file.exists(grunt.config.get('config.buildPaths.html') + '/index.php')) {
-    tasksDefault.push('newer:drushmake:default');
-  }
-  else {
-    tasksDefault.push('drushmake:default');
+  // Process .make files if configured.
+  if (grunt.config.get('config.srcPaths.make')) {
+    // If build/html exists, but is empty, skip the newer check.
+    // This facilitates situations where the build/html is generated as a mounted
+    // directory point with a newer timestamp than the Drush Makefiles.
+    //
+    // We do not use the grunt-newer .cache with drushmake so skipping newer for
+    // any one run does not impact later behavior.
+    if (grunt.file.exists(grunt.config.get('config.buildPaths.html') + '/index.php')) {
+      tasksDefault.push('newer:drushmake:default');
+    } else {
+      tasksDefault.push('drushmake:default');
+    }
   }
 
   // Wire up the generated docroot to our custom code.
   tasksDefault.push('scaffold');
 
-  if (grunt.config.get(['composer', 'install'])) {
+  if (grunt.file.exists('./composer.lock') && grunt.config.get(['composer', 'install'])) {
+    // Manually run `composer drupal-scaffold` since this is only automatically run on update.
+    tasksDefault.unshift('composer:drupal-scaffold');
+    // Run `composer install` if there is already a lock file. Updates should be explicit once this file exists.
     tasksDefault.unshift('composer:install');
+  } else if (grunt.config.get(['composer', 'update'])) {
+    // Run `composer update` if no lock file exists. This forces `composer drupal-scaffold` to run.
+    tasksDefault.unshift('composer:update');
   }
-  if (grunt.task.exists('bundle-install')) {
-    tasksDefault.unshift('bundle-install');
-  }
+
   if (grunt.task.exists('compile-theme')) {
     tasksDefault.push('compile-theme');
   }