+++ tasks/clean.js
@@ -1,5 +1,4 @@
 module.exports = function(grunt) {
-
   /**
    * Define "clean" tasks to remove files and directories.
    *
@@ -25,7 +24,7 @@ module.exports = function(grunt) {
   });
 
   Help.add({
-    task: 'clean',
+    task: 'clean:default',
     group: 'Utilities',
     description: 'Use "clean" to remove the build output directory.'
   });