+++ tasks/package.js
@@ -1,5 +1,4 @@
 module.exports = function(grunt) {
-
   /**
    * Define "package" tasks.
    *
@@ -8,17 +7,20 @@ module.exports = function(grunt) {
    */
 
   var Help = require('../lib/help')(grunt);
+  var path = require('path');
 
   grunt.registerTask('package', 'Package the operational codebase for deployment. Use package:compress to create an archive.', function() {
-    var path = require('path');
     grunt.loadNpmTasks('grunt-contrib-copy');
 
-    var config = grunt.config.get('config.packages'),
-      srcFiles = ['**', '!**/.gitkeep'].concat((config && config.srcFiles && config.srcFiles.length) ? config.srcFiles : '**'),
-      projFiles = (config && config.projFiles && config.projFiles.length) ? config.projFiles : [];
+    var config = grunt.config.get('config.packages');
+    var srcFiles = ['**', '!**/.gitkeep'].concat((config && config.srcFiles && config.srcFiles.length) ? config.srcFiles : '**');
+    var projFiles = (config && config.projFiles && config.projFiles.length) ? config.projFiles : [];
+    console.log(projFiles);
 
-    var destPath = grunt.config.get('config.buildPaths.package') + '/package';
-    var tasks = []
+    // Look for a package target spec, build destination path.
+    var packageName = grunt.option('name') || config.name || 'package';
+    var destPath = grunt.config.get('config.buildPaths.packages') + '/' + packageName;
+    var tasks = [];
 
     grunt.config('copy.package', {
       files: [
@@ -26,14 +28,14 @@ module.exports = function(grunt) {
           expand: true,
           cwd: '<%= config.buildPaths.html %>',
           src: srcFiles,
-          dest: destPath + (grunt.config.get('config.packages.dest.docroot') || ''),
+          dest: path.resolve(destPath, grunt.config.get('config.packages.dest.docroot') || ''),
           dot: true,
           follow: true
         },
         {
           expand: true,
           src: projFiles,
-          dest: destPath + (grunt.config.get('config.packages.dest.devResources') || ''),
+          dest: path.resolve(destPath, grunt.config.get('config.packages.dest.devResources') || ''),
           dot: true,
           follow: true
         }
@@ -44,9 +46,25 @@ module.exports = function(grunt) {
       }
     });
 
+    grunt.config.set('clean.packages', [destPath]);
+
+    tasks.push('clean:packages');
     tasks.push('copy:package');
 
-    if (this.args[0] && this.args[0] == 'compress') {
+    // If the `composer.json` file is being packaged, rebuild composer dependencies without dev.
+    if (projFiles.find(function(pattern) {
+      return pattern.startsWith('composer');
+    })) {
+      grunt.config(['composer'], {
+        options: {
+          flags: ['no-dev'],
+          cwd: destPath
+        }
+      });
+      tasks.push('composer:install');
+    }
+
+    if (this.args[0] && this.args[0] === 'compress') {
       grunt.loadNpmTasks('grunt-contrib-compress');
       grunt.config('compress.package', {
         options: {
@@ -58,8 +76,8 @@ module.exports = function(grunt) {
           {
             expand: true,
             dot: true,
-            cwd: grunt.config.get('config.buildPaths.package'),
-            src: 'package/**',
+            cwd: grunt.config.get('config.buildPaths.packages') + '/' + packageName,
+            src: ['**']
           }
         ]
       });