+++ tasks/serve.js
@@ -1,5 +1,4 @@
 module.exports = function(grunt) {
-
   /**
    * Define tasks to quickly get Drupal running.
    *
@@ -13,23 +12,27 @@ module.exports = function(grunt) {
    */
   grunt.loadNpmTasks('grunt-drush');
 
-  var Help = require('../lib/help')(grunt),
-    Drupal = require('../lib/drupal')(grunt);
+  var Help = require('../lib/help')(grunt);
+  var Drupal = require('../lib/drupal')(grunt);
 
-  var path = require('path'),
-    _ = require('lodash');
+  var path = require('path');
+  var _ = require('lodash');
 
   // If no path is configured for Drush, fallback to the system path.
   var cmd = {cmd: Drupal.drushPath()};
 
   // `config.serve.profile` is deprecated.
   if (grunt.config('config.serve.profile')) {
-    grunt.log.warn('The `serve.profile` parameter is deprecated and will be removed in future versions. Use `project.profile` instead.');
+    grunt.log.warn('The `serve.profile` parameter is deprecated and will be ' +
+      'removed in future versions. Use `project.profile` instead.');
   }
-  var profile = grunt.config('config.project.profile') || grunt.config('config.serve.profile') || 'standard';
+  var profile = grunt.config('config.project.profile') ||
+    grunt.config('config.serve.profile') || 'standard';
 
   grunt.config(['drush', 'liteinstall'], {
-    args: ['site-install', '-yv', profile, '--db-url=sqlite:/' + path.join(path.resolve(grunt.config('config.buildPaths.build')), 'drupal.sqlite')],
+    args: ['site-install', '-yv', profile, '--db-url=sqlite:/' +
+      path.join(path.resolve(grunt.config('config.buildPaths.build')),
+      'drupal.sqlite')],
     options: _.extend({
       cwd: '<%= config.buildPaths.html %>'
     }, cmd)
@@ -42,44 +45,46 @@ module.exports = function(grunt) {
     }, cmd)
   });
 
-  grunt.registerTask('serve', 'Run Drupal using the PHP built-in webserver. serve:demo to run without watch tasks and serve:test for grunt-drupal-tasks development.', function() {
-
-    if (this.args[0] != 'test') {
-      // Unlike the test version, this one allows port overriding and opens the site in a new tab.
-      grunt.config('drush.serve.args', ['runserver', ':' + (grunt.config('config.serve.port') || 8080) + '/' ]);
-    }
-
-    if (this.args.length) {
-      grunt.task.run('drush:serve');
-    }
-    else {
-      grunt.loadNpmTasks('grunt-concurrent');
-
-      // Allow overriding the tasks run concurrently to the Drupal server.
-      var serveTasks = grunt.config('serve.concurrent');
-      if (!serveTasks) {
-        serveTasks = ['watch-test'];
-        if (grunt.task.exists('watch-theme')) {
-          serveTasks.push('watch-theme');
-        }
+  grunt.registerTask('serve', 'Run Drupal using the PHP built-in webserver. ' +
+    'Use serve:demo to run without watch tasks. Use serve:test for G-D-T ' +
+    'development.', function() {
+      if (this.args[0] !== 'test') {
+        // Unlike the test version, this one allows port overriding and opens
+        // the site in a new tab.
+        grunt.config('drush.serve.args', ['runserver', ':' +
+          (grunt.config('config.serve.port') || 8080) + '/']);
       }
-      serveTasks.push('drush:serve');
 
-      grunt.config('concurrent.serve', {
-        tasks: serveTasks,
-        options: {
-          logConcurrentOutput: true
+      if (this.args.length) {
+        grunt.task.run('drush:serve');
+      } else {
+        grunt.loadNpmTasks('grunt-concurrent');
+
+        // Allow overriding the tasks run concurrently to the Drupal server.
+        var serveTasks = grunt.config('serve.concurrent');
+        if (!serveTasks) {
+          serveTasks = ['watch-test'];
+          if (grunt.task.exists('watch-theme')) {
+            serveTasks.push('watch-theme');
+          }
         }
-      });
+        serveTasks.push('drush:serve');
 
-      grunt.task.run('concurrent:serve');
-    }
-  });
+        grunt.config('concurrent.serve', {
+          tasks: serveTasks,
+          options: {
+            logConcurrentOutput: true
+          }
+        });
+
+        grunt.task.run('concurrent:serve');
+      }
+    });
 
   Help.add([
     {
       task: 'serve',
       group: 'Operations'
-    },
+    }
   ]);
 };