+++ test/test.sh
@@ -20,13 +20,13 @@ run_tests () {
   mocha node_modules/grunt-drupal-tasks/test/library.js
 }
 
-# Drupal 7
-export GDT_DRUPAL_CORE="7" GDT_TEST_URL="http://127.0.0.1:8080/misc/drupal.js"
+# Drupal 8
+export GDT_DRUPAL_CORE="8" GDT_TEST_URL="http://127.0.0.1:8080/core/misc/drupal.js"
 run_tests
 
 # Rinse and repeat
 cd ../..
 
-# Drupal 8
-export GDT_DRUPAL_CORE="8" GDT_TEST_URL="http://127.0.0.1:8080/core/misc/drupal.js"
+# Drupal 7
+export GDT_DRUPAL_CORE="7" GDT_TEST_URL="http://127.0.0.1:8080/misc/drupal.js"
 run_tests