+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # CHANGELOG
 
+## v1.0.0-pre
+
+- Dropped support for Node.js v0.12 and earlier and PHP v5.4 and earlier.
+- Added support for a Composer build process for Drupal 8.
+- Removed built-in support for Compass theme compilation and Ruby bundle
+installation.
+- Added configuration for the Behat binary path. Add default configuration for
+the paths of the phpcs, phpmd, and Drush binaries. Changed configuration key
+from `drush.cmd` to `drush.path` for consistency.
+- Minor code refactoring and adopting code style standard.
+- Updated dependencies.
+
+### Upgrade Notes
+
+- Node.js v4 or later is required. Grunt Drupal Tasks is now compatible with
+Node.js v5.
+- PHP v5.5 or later is required.
+- Use of the built-in Compass theme compilation steps must be replaced by custom
+handling at the project- or theme-level.
+- Change the Gruntconfig.json configuration key `cmd` under `drush` to `path`,
+or if using the default path of `vendor/bin/drush`, remove the setting entirely.
+
 ## v0.11.1 [2016/04/17]
 
 - Fixed `grunt package` regression that custom code from the packaged outputs.